+++ src/charts/helpers/timeAxis.js
@@ -7,7 +7,7 @@ define(function(require) {
     const {
         axisTimeCombinations,
         timeBenchmarks
-    } = require('./constants');
+    } = require('../../helpers/constants');
 
 
     const singleTickWidth = 20;
