+++ src/charts/stacked-area.js
@@ -19,7 +19,7 @@ define(function(require){
         getXAxisSettings,
         getLocaleDateFormatter
     } = require('./helpers/timeAxis');
-    const {axisTimeCombinations, curveMap} = require('./helpers/constants');
+    const {axisTimeCombinations, curveMap} = require('../helpers/constants');
     const {
         formatIntegerValue,
         formatDecimalValue
